requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
flask-mail==0.9.1
python-dotenv==1.0.0
cryptography==41.0.4
pytest==7.4.2
pydantic==2.4.2
python-magic==0.4.27 